Feature shriram hms

+++ src/main/java/com/hospital/management/service/InsuranceCompService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface InsuranceCompService {
 
     InsuranceComp saveInsuranceComp(InsuranceComp insuranceComp);
-    InsuranceComp updateInsuranceComp(InsuranceComp insuranceComp, Integer insComId);
+    InsuranceComp updateInsuranceComp(InsuranceComp insuranceComp, Long insComId);
     List<InsuranceComp> insuranceCompList();
-    InsuranceComp findInsuranceCompByInsComId(Integer insComId);
-    String deleteInsuranceCompByInsComId(Integer insComId);
+    InsuranceComp findInsuranceCompByInsComId(Long insComId);
+    String deleteInsuranceCompByInsComId(Long insComId);
 
 }
